backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-groq
langgraph
faiss-cpu
python-dotenv